Remove parent child hierarchy
