+++ src/main/java/com/uid2/shared/secure/AzureCCCoreAttestationService.java
@@ -17,15 +17,77 @@ import java.util.Set;
 
 // CC stands for Confidential Container
 @Slf4j
-public class AzureCCCoreAttestationService extends AzureCCCoreAttestationServiceBase {
-    private static final String AZURE_CC_PROTOCOL = "azure-cc";
+public class AzureCCCoreAttestationService implements ICoreAttestationService {
 
-    public AzureCCCoreAttestationService(String maaServerBaseUrl, String attestationUrl) {
-        super(new MaaTokenSignatureValidator(maaServerBaseUrl), new PolicyValidator(attestationUrl), AZURE_CC_PROTOCOL);
+    protected final Set<String> allowedEnclaveIds = new HashSet<>();
+
+    protected final IMaaTokenSignatureValidator tokenSignatureValidator;
+
+    protected final IPolicyValidator policyValidator;
+
+    protected final String azureCcProtocol;
+
+    public AzureCCCoreAttestationService(String maaServerBaseUrl, String attestationUrl, String azureCcProtocol) {
+        this(new MaaTokenSignatureValidator(maaServerBaseUrl), new PolicyValidator(attestationUrl), azureCcProtocol);
     }
 
     // used in UT
-    protected AzureCCCoreAttestationService(IMaaTokenSignatureValidator tokenSignatureValidator, IPolicyValidator policyValidator) {
-        super(tokenSignatureValidator, policyValidator, AZURE_CC_PROTOCOL);
+    protected AzureCCCoreAttestationService(IMaaTokenSignatureValidator tokenSignatureValidator, IPolicyValidator policyValidator, String azureCcProtocol) {
+        this.tokenSignatureValidator = tokenSignatureValidator;
+        this.policyValidator = policyValidator;
+        this.azureCcProtocol = azureCcProtocol;
+    }
+
+    @Override
+    public void attest(byte[] attestationRequest, byte[] publicKey, Handler<AsyncResult<AttestationResult>> handler) {
+        try {
+            var tokenString = new String(attestationRequest, StandardCharsets.US_ASCII);
+
+            log.debug("Validating signature...");
+            var tokenPayload = tokenSignatureValidator.validate(tokenString, azureCcProtocol);
+
+            log.debug("Validating policy...");
+            var encodedPublicKey = Utils.toBase64String(publicKey);
+
+            var enclaveId = policyValidator.validate(tokenPayload, encodedPublicKey);
+
+            if (allowedEnclaveIds.contains(enclaveId)) {
+                log.info("Successfully attested " + azureCcProtocol + " against registered enclaves, enclave id: " + enclaveId);
+                handler.handle(Future.succeededFuture(new AttestationResult(publicKey, enclaveId)));
+            } else {
+                log.warn("Got unsupported " + azureCcProtocol + " enclave id: " + enclaveId);
+                handler.handle(Future.succeededFuture(new AttestationResult(AttestationFailure.FORBIDDEN_ENCLAVE)));
+            }
+        }
+        catch (AttestationClientException ace){
+            handler.handle(Future.succeededFuture(new AttestationResult(ace)));
+        } catch (AttestationException ae) {
+            handler.handle(Future.failedFuture(ae));
+        } catch (Exception ex) {
+            handler.handle(Future.failedFuture(new AttestationException(ex)));
+        }
+    };
+
+    @Override
+    public void registerEnclave(String encodedIdentifier) throws AttestationException {
+        try {
+            allowedEnclaveIds.add(encodedIdentifier);
+        } catch (Exception e) {
+            throw new AttestationException(e);
+        }
+    }
+
+    @Override
+    public void unregisterEnclave(String encodedIdentifier) throws AttestationException {
+        try {
+            allowedEnclaveIds.remove(encodedIdentifier);
+        } catch (Exception e) {
+            throw new AttestationException(e);
+        }
+    }
+
+    @Override
+    public Collection<String> getEnclaveAllowlist() {
+        return allowedEnclaveIds;
     }
 }
