+++ src/test/java/com/uid2/shared/secure/AzureCCCoreAttestationServiceTest.java
@@ -9,6 +9,8 @@ import io.vertx.core.Handler;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.ValueSource;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.mockito.junit.jupiter.MockitoSettings;
@@ -59,9 +61,10 @@ class AzureCCCoreAttestationServiceTest {
         when(alwaysPassPolicyValidator.validate(any(), any())).thenReturn(ENCLAVE_ID);
     }
 
-    @Test
-    public void testHappyPath() throws AttestationException {
-        var provider = new AzureCCCoreAttestationService(alwaysPassTokenValidator, alwaysPassPolicyValidator);
+    @ParameterizedTest
+    @ValueSource(strings = {"azure-cc", "azure-cc-aks"})
+    public void testHappyPath(String azureProtocol) throws AttestationException {
+        var provider = new AzureCCCoreAttestationService(alwaysPassTokenValidator, alwaysPassPolicyValidator, azureProtocol);
         provider.registerEnclave(ENCLAVE_ID);
         attest(provider, ar -> {
             assertTrue(ar.succeeded());
@@ -69,11 +72,12 @@ class AzureCCCoreAttestationServiceTest {
         });
     }
 
-    @Test
-    public void testSignatureCheckFailed_ClientError() throws AttestationException {
+    @ParameterizedTest
+    @ValueSource(strings = {"azure-cc", "azure-cc-aks"})
+    public void testSignatureCheckFailed_ClientError(String azureProtocol) throws AttestationException {
         var errorStr = "token signature validation failed";
         when(alwaysFailTokenValidator.validate(any(), any())).thenThrow(new AttestationClientException(errorStr, AttestationFailure.BAD_PAYLOAD));
-        var provider = new AzureCCCoreAttestationService(alwaysFailTokenValidator, alwaysPassPolicyValidator);
+        var provider = new AzureCCCoreAttestationService(alwaysFailTokenValidator, alwaysPassPolicyValidator, azureProtocol);
         provider.registerEnclave(ENCLAVE_ID);
         attest(provider, ar -> {
             assertTrue(ar.succeeded());
@@ -82,10 +86,11 @@ class AzureCCCoreAttestationServiceTest {
         });
     }
 
-    @Test
-    public void testSignatureCheckFailed_ServerError() throws AttestationException {
+    @ParameterizedTest
+    @ValueSource(strings = {"azure-cc", "azure-cc-aks"})
+    public void testSignatureCheckFailed_ServerError(String azureProtocol) throws AttestationException {
         when(alwaysFailTokenValidator.validate(any(), any())).thenThrow(new AttestationException("unknown server error"));
-        var provider = new AzureCCCoreAttestationService(alwaysFailTokenValidator, alwaysPassPolicyValidator);
+        var provider = new AzureCCCoreAttestationService(alwaysFailTokenValidator, alwaysPassPolicyValidator, azureProtocol);
         provider.registerEnclave(ENCLAVE_ID);
         attest(provider, ar -> {
             assertFalse(ar.succeeded());
@@ -93,11 +98,12 @@ class AzureCCCoreAttestationServiceTest {
         });
     }
 
-    @Test
-    public void testPolicyCheckSuccess_ClientError() throws AttestationException {
+    @ParameterizedTest
+    @ValueSource(strings = {"azure-cc", "azure-cc-aks"})
+    public void testPolicyCheckSuccess_ClientError(String azureProtocol) throws AttestationException {
         var errorStr = "policy validation failed";
         when(alwaysFailPolicyValidator.validate(any(), any())).thenThrow(new AttestationClientException(errorStr, AttestationFailure.BAD_PAYLOAD));
-        var provider = new AzureCCCoreAttestationService(alwaysFailTokenValidator, alwaysFailPolicyValidator);
+        var provider = new AzureCCCoreAttestationService(alwaysFailTokenValidator, alwaysFailPolicyValidator, azureProtocol);
         provider.registerEnclave(ENCLAVE_ID);
         attest(provider, ar -> {
             assertTrue(ar.succeeded());
@@ -106,10 +112,11 @@ class AzureCCCoreAttestationServiceTest {
         });
     }
 
-    @Test
-    public void testPolicyCheckFailed_ServerError() throws AttestationException {
+    @ParameterizedTest
+    @ValueSource(strings = {"azure-cc", "azure-cc-aks"})
+    public void testPolicyCheckFailed_ServerError(String azureProtocol) throws AttestationException {
         when(alwaysFailPolicyValidator.validate(any(), any())).thenThrow(new AttestationException("unknown server error"));
-        var provider = new AzureCCCoreAttestationService(alwaysFailTokenValidator, alwaysFailPolicyValidator);
+        var provider = new AzureCCCoreAttestationService(alwaysFailTokenValidator, alwaysFailPolicyValidator, azureProtocol);
         provider.registerEnclave(ENCLAVE_ID);
         attest(provider, ar -> {
             assertFalse(ar.succeeded());
@@ -117,9 +124,10 @@ class AzureCCCoreAttestationServiceTest {
         });
     }
 
-    @Test
-    public void testEnclaveNotRegistered() throws AttestationException {
-        var provider = new AzureCCCoreAttestationService(alwaysFailTokenValidator, alwaysPassPolicyValidator);
+    @ParameterizedTest
+    @ValueSource(strings = {"azure-cc", "azure-cc-aks"})
+    public void testEnclaveNotRegistered(String azureProtocol) throws AttestationException {
+        var provider = new AzureCCCoreAttestationService(alwaysFailTokenValidator, alwaysPassPolicyValidator, azureProtocol);
         attest(provider, ar -> {
             assertTrue(ar.succeeded());
             assertFalse(ar.result().isSuccess());
